requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

cryptography==44.0.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pyotp==2.9.0
python-dotenv==1.0.1
